Auth.js - Autenticacion de proteccion de rutas

+++ src/app/api/todos/route.ts
@@ -1,3 +1,4 @@
+import { getUserSessionServer } from '@/auth/actions/auth-actions';
 import prisma from '@/lib/prisma'
 import { NextResponse, NextRequest } from 'next/server'
 import * as yup from 'yup';
@@ -30,10 +31,18 @@ const postSchema = yup.object({
 
 export async function POST(request: Request) {
 
+    const user = await getUserSessionServer();
+
+    if (!user) {
+        return NextResponse.json("No autorizado", { status: 401 });
+
+    }
+
+
     try {
         const { completed, description } = await postSchema.validate(await request.json());
 
-        const todo = await prisma.todo.create({ data: { completed, description } });
+        const todo = await prisma.todo.create({ data: { completed, description, userId: user.id } });
 
         return NextResponse.json(todo);
 
@@ -44,10 +53,15 @@ export async function POST(request: Request) {
 }
 export async function DELETE(request: Request) {
 
+    const user = await getUserSessionServer();
+
+    if (!user) {
+        return NextResponse.json("No autorizado", { status: 401 });
+
+    }
+
     try {
-        const todo = await prisma.todo.deleteMany({
-            where: { completed: true }
-        });
+        const todo = await prisma.todo.deleteMany({ where: { completed: true, userId: user.id } });
 
         return NextResponse.json("Borrados");
